server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bpython>=0.25",
    "celery[auth,librabbitmq,msgpack,redis]>=5.5.3",
    "django>=5.2.1",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-money>=3.5.4",
    "django-phonenumber-field[phonenumberslite]>=8.1.0",
    "django-redis>=5.4.0",
    "django-silk>=5.3.2",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt[crypto]>=5.5.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "markdown>=3.8",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.9",
    "pyclean>=3.1.0",
    "ruff>=0.11.12",
    "uvicorn>=0.34.3",
]

[tool.ruff]
line-length = 89
exclude = ["migrations", "staticfiles"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 20
